# api/openapi.yaml
openapi: 3.1.0
info:
  title: Chemistry Utility API
  version: 1.0.0
  description: |
    REST API for the Chemistry Utility application.

    Provides chemistry calculators (molar mass, equation balancing, stoichiometry,
    gas laws, thermodynamics, kinetics, electrochemistry, quantum, solutions),
    compound search, calculation history, workspace collaboration, API key
    management, analytics, and plugin administration.

    ## Authentication
    Most endpoints require a JWT bearer token obtained via `POST /auth/login`,
    `POST /auth/register`, or OAuth. Include the token in the `Authorization`
    header as `Bearer <access_token>`.

    ## Roles
    - `student` (default for new registrations): can use calculators and manage own data.
    - `researcher`: student permissions plus API key management.
    - `educator`: same as researcher (granted by admins).
    - `admin`: full access including analytics and plugin management.

    ## Errors
    All error responses use RFC 7807 Problem Details (`application/problem+json`).
  contact:
    name: Chemistry Utility
    url: https://chemistry-utility.dev
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:6005/api/v1
    description: Local development server
  - url: https://api.chemistry-utility.dev/api/v1
    description: Production server
security:
  - BearerAuth: []
...
paths:
  /auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: Creates a new user account with the `student` role and returns auth tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '409':
          $ref: '#/components/responses/ConflictError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in with email and password
      description: Authenticates a user and returns access/refresh tokens.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Exchanges a valid refresh token for a new token pair.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshRequest'
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema:
                type: object
                properties:
                  tokens:
                    $ref: '#/components/schemas/TokenPair'
                required: [tokens]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request a password reset link
      description: |
        Triggers a password reset flow for the given email. Always returns success
        to prevent email enumeration, regardless of whether the email exists.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Reset email sent (if account exists)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: if the email exists, a reset link has been sent
                required: [message]
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password using a reset token
      description: Sets a new password for a user given a valid reset token.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Password reset successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: password has been reset
                required: [message]
        '400':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/github:
    get:
      tags: [Auth]
      summary: Begin GitHub OAuth flow
      description: Redirects to GitHub's OAuth authorization page. Returns 404 if GitHub OAuth is not configured.
      security: []
      responses:
        '302':
          description: Redirect to GitHub authorization URL
          headers:
            Location:
              schema:
                type: string
                format: uri
        '404':
          $ref: '#/components/responses/NotFoundError'
  /auth/github/callback:
    get:
      tags: [Auth]
      summary: GitHub OAuth callback
      description: Handles the OAuth callback from GitHub, exchanges the code, and returns auth tokens.
      security: []
      parameters:
        - name: code
          in: query
          required: true
          description: Authorization code returned by GitHub
          schema:
            type: string
      responses:
        '200':
          description: OAuth login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /auth/google:
    get:
      tags: [Auth]
      summary: Begin Google OAuth flow
      description: Redirects to Google's OAuth authorization page. Returns 404 if Google OAuth is not configured.
      security: []
      responses:
        '302':
          description: Redirect to Google authorization URL
          headers:
            Location:
              schema:
                type: string
                format: uri
        '404':
          $ref: '#/components/responses/NotFoundError'
  /auth/google/callback:
    get:
      tags: [Auth]
      summary: Google OAuth callback
      description: Handles the OAuth callback from Google, exchanges the code, and returns auth tokens.
      security: []
      parameters:
        - name: code
          in: query
          required: true
          description: Authorization code returned by Google
          schema:
            type: string
      responses:
        '200':
          description: OAuth login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /users/me:
    get:
      tags: [Users]
      summary: Get the current user's profile
      description: Returns the authenticated user's public profile.
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUser'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Users]
      summary: Update the current user's profile
      description: Updates the authenticated user's name and/or email.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: Updated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PublicUser'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /calculators:
    get:
      tags: [Calculators]
      summary: List available calculator types
      description: Returns the sorted list of all registered calculator type identifiers.
      security: []
      responses:
        '200':
          description: List of calculator types
          content:
            application/json:
              schema:
                type: object
                properties:
                  calculators:
                    type: array
                    items:
                      type: string
                    example: [arrhenius, bond-type, buffer-solution, cell-potential]
                required: [calculators]
        '429':
          $ref: '#/components/responses/RateLimitError'
  /calculators/{type}:
    post:
      tags: [Calculators]
      summary: Run a calculator
      description: |
        Executes the calculator identified by `type` with the provided input map.
        The shape of `inputs` depends on the calculator type; see the developer
        guide for per-calculator input schemas. Returns a `CalculationResult`.
      security: []
      parameters:
        - $ref: '#/components/parameters/CalculatorType'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Calculator-specific input map. See developer guide for details.
              additionalProperties: true
            example:
              formula: H2O
      responses:
        '200':
          description: Calculator result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculationResult'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /calculations:
    get:
      tags: [Calculations]
      summary: List the current user's calculations
      description: Returns a paginated list of the authenticated user's saved calculations.
      parameters:
        - $ref: '#/components/parameters/PageQuery'
        - $ref: '#/components/parameters/LimitQuery'
      responses:
        '200':
          description: Paginated list of calculations
          content:
            application/json:
              schema:
                type: object
                properties:
                  calculations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Calculation'
                  page:
                    type: integer
                    minimum: 1
                    example: 1
                  limit:
                    type: integer
                    minimum: 1
                    maximum: 100
                    example: 20
                required: [calculations, page, limit]
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /calculations/{id}:
    get:
      tags: [Calculations]
      summary: Get a calculation by ID
      description: Returns a specific calculation. Only the owner or an admin may view it.
      parameters:
        - $ref: '#/components/parameters/CalculationIdPath'
      responses:
        '200':
          description: The calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calculation'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Calculations]
      summary: Delete a calculation
      description: Permanently deletes a calculation. Only the owner or an admin may delete it.
      parameters:
        - $ref: '#/components/parameters/CalculationIdPath'
      responses:
        '204':
          description: Calculation deleted successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /calculations/{id}/annotate:
    post:
      tags: [Calculations]
      summary: Annotate a calculation
      description: Sets or updates the annotation text on a calculation. Only the owner or an admin may annotate.
      parameters:
        - $ref: '#/components/parameters/CalculationIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                annotation:
                  type: string
                  description: Free-form annotation text
              required: [annotation]
      responses:
        '200':
          description: Updated calculation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Calculation'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /calculations/{id}/star:
    post:
      tags: [Calculations]
      summary: Toggle star on a calculation
      description: Toggles the starred flag on a calculation and returns the new state. Only the owner or an admin may star.
      parameters:
        - $ref: '#/components/parameters/CalculationIdPath'
      responses:
        '200':
          description: New starred state
          content:
            application/json:
              schema:
                type: object
                properties:
                  starred:
                    type: boolean
                required: [starred]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /workspaces:
    get:
      tags: [Workspaces]
      summary: List the current user's workspaces
      description: Returns all workspaces the authenticated user belongs to.
      responses:
        '200':
          description: List of workspaces
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
                required: [workspaces]
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Workspaces]
      summary: Create a workspace
      description: Creates a new workspace and adds the creator as the `owner` member.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWorkspaceRequest'
      responses:
        '201':
          description: Workspace created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /workspaces/{id}:
    get:
      tags: [Workspaces]
      summary: Get a workspace by ID
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
      responses:
        '200':
          description: The workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    patch:
      tags: [Workspaces]
      summary: Update a workspace
      description: Updates the name and/or description. Only the owner or an admin may update.
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateWorkspaceRequest'
      responses:
        '200':
          description: Updated workspace
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Workspace'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags: [Workspaces]
      summary: Delete a workspace
      description: Permanently deletes a workspace. Only the owner or an admin may delete.
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
      responses:
        '204':
          description: Workspace deleted successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /workspaces/{id}/members:
    post:
      tags: [Workspaces]
      summary: Add a member to a workspace
      description: Adds a user to the workspace with the given role. Only the owner or an admin may add members.
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddMemberRequest'
      responses:
        '201':
          description: Member added
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceMember'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /workspaces/{id}/members/{userId}:
    delete:
      tags: [Workspaces]
      summary: Remove a member from a workspace
      description: Removes a user from the workspace. Only the owner or an admin may remove members.
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
        - $ref: '#/components/parameters/MemberUserIdPath'
      responses:
        '204':
          description: Member removed successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/InternalError'
  /workspaces/{id}/calculations:
    get:
      tags: [Workspaces]
      summary: List calculations in a workspace
      description: Returns all calculations associated with the given workspace.
      parameters:
        - $ref: '#/components/parameters/WorkspaceIdPath'
      responses:
        '200':
          description: List of workspace calculations
          content:
            application/json:
              schema:
                type: object
                properties:
                  calculations:
                    type: array
                    items:
                      $ref: '#/components/schemas/Calculation'
                required: [calculations]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/InternalError'
  /compounds:
    get:
      tags: [Compounds]
      summary: Search compounds
      description: Searches compounds by name, formula, or other indexed fields.
      security: []
      parameters:
        - name: q
          in: query
          required: true
          description: Search query string
          schema:
            type: string
            minLength: 1
        - $ref: '#/components/parameters/LimitQuery'
        - $ref: '#/components/parameters/OffsetQuery'
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  compounds:
                    type: array
                    items:
                      $ref: '#/components/schemas/Compound'
                  query:
                    type: string
                required: [compounds, query]
        '400':
          $ref: '#/components/responses/ValidationError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /compounds/{id}:
    get:
      tags: [Compounds]
      summary: Get a compound by ID
      security: []
      parameters:
        - $ref: '#/components/parameters/CompoundIdPath'
      responses:
        '200':
          description: The compound
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Compound'
        '400':
          $ref: '#/components/responses/ValidationError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '429':
          $ref: '#/components/responses/RateLimitError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api-keys:
    get:
      tags: [APIKeys]
      summary: List the current user's API keys
      description: Returns all API keys owned by the authenticated user. Requires `researcher`, `educator`, or `admin` role.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: List of API keys (without raw key values)
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_keys:
                    type: array
                    items:
                      $ref: '#/components/schemas/APIKey'
                required: [api_keys]
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [APIKeys]
      summary: Create an API key
      description: |
        Creates a new API key for the authenticated user. The raw key value is
        only returned once in the response and cannot be retrieved again.
        Requires `researcher`, `educator`, or `admin` role.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Human-readable label for the key
              required: [name]
      responses:
        '201':
          description: API key created (includes raw key value)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIKey'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalError'
  /api-keys/{id}:
    delete:
      tags: [APIKeys]
      summary: Delete an API key
      description: Permanently deletes an API key by ID. Requires `researcher`, `educator`, or `admin` role.
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/APIKeyIdPath'
      responses:
        '204':
          description: API key deleted successfully
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalError'
  /analytics/overview:
    get:
      tags: [Analytics]
      summary: Platform overview statistics
      description: Returns aggregated platform statistics. Admin only.
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Overview statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  overview:
                    type: object
                    additionalProperties: true
                    description: Aggregated platform statistics
                required: [overview]
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token obtained from /auth/login, /auth/register, or OAuth callback.
...
  responses:
...
    InternalError:
      description: Unexpected server error (RFC 7807).
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
          example: alice@example.com
        password:
          type: string
          format: password
          minLength: 8
          example: super-secret-123
        name:
          type: string
          example: Alice Doe
      required: [email, password, name]
...